+++ bldrx/__init__.py
@@ -1,3 +1,3 @@
 """bldrx package"""
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"